project_part_2/requirements.txt
flask~=2.0.2
requests~=2.26.0
flask_socketio
kafka-python